* origin/caf/github-qca/master:
  NAN: Send IPv6 address, port and protocol in TLV format
  Add wifi hal initialization in location tests
  Add more general function to initialize wifi hal
  dev_exec_action: Clear domain{,_suffix}_match on UOSC
  ap_config_commit: Fix hostapd configuration for FT-EAP only AKM list
  dev_start/stop_test: Remove vendor name default
  dev_exec_action: Allow UOSP to be disabled
  HS 2.0: Use e_loop to start hs20-osu-client with more privilege

CRs-Fixed: 2467808

+++ sta.c
@@ -7501,6 +7501,10 @@ static enum sigma_cmd_result cmd_sta_reset_default(struct sigma_dut *dut,
 		nan_cmd_sta_reset_default(dut, conn, cmd);
 #endif /* ANDROID_NAN */
 
+	if (dut->program == PROGRAM_LOC &&
+	    lowi_cmd_sta_reset_default(dut, conn, cmd) < 0)
+		return ERROR_SEND_STATUS;
+
 	if (dut->program == PROGRAM_HS2_R2 || dut->program == PROGRAM_HS2_R3) {
 		unlink("SP/wi-fi.org/pps.xml");
 		if (system("rm -r SP/*") != 0) {
